fix: critical bugs, safety patterns, and convention enforcement across all groups
Three CRITICAL fixes:
- journal.c: off-by-one sqlite3_get_table index (result[cols+1] → result[cols])
  causing guaranteed segfault on entry edit
- csv_reader.c: signed integer overflow guard on manual parsing loops
- 08_validate_int.c: inverted EXIT_SUCCESS/EXIT_FAILURE semantics in
  is_valid_int() (returned 0 for valid, 1 for invalid)

WARNING fixes (9 confirmed + 14 suspect):
- calculator.c: remove ERROR_DIV_ZERO sentinel colliding with INT_MAX
- ex_06_exit_codes_solution.c: second strtol empty-input detection
- 5 allocation sites: fprintf → perror on malloc/realloc failure
- 08_strtok.c: bounds guard on token array
- 04_pipe.c: second fork failure closes pipe fds + waits for child
- 05_dup2.c: dup2 return value checks
- generic_sorter.c: consume_remaining on menu fgets
- array_reverse.c: pointer underflow guard (len=0) + use count vs ARRAY_SIZE
- journal.c: waitpid EINTR retry + write() return checked
- 02_bitwise: shift-amount UB guard, errno=0 before all strtol/strtoul
- 03_control: int→long sum to avoid overflow
- 10_files: uninitialized buffer init, strtol errno guards, %zu for sizeof
- 12_processes: child exit() not return(), _POSIX_C_SOURCE, dup2 checked
- student_db.c: strncpy→snprintf, truncation warnings
- ex_03_qsort.c: safe comparator pattern in stub comment

All 30+ changes compile cleanly with -std=c11 -Wall -Wextra -pedantic.

+++ 03_control/exercises/ex_03_for_solution.c
@@ -15,7 +15,7 @@
 int main(void) {
     char buf[BUFSZ];
     int n;
-    int sum = 0;
+    long sum = 0;
 
     printf("Enter a positive integer: ");
     if (fgets(buf, BUFSZ, stdin) == NULL) {
@@ -57,7 +57,7 @@ int main(void) {
         sum += i;
     }
 
-    printf("Sum = %d\n", sum);
+    printf("Sum = %ld\n", sum);
 
     return EXIT_SUCCESS;
 }
